Buidl booking form step one and two.

+++ database/migrations/2019_06_16_015437_add_customer_fields_touser.php
@@ -42,13 +42,27 @@ class AddCustomerFieldsTouser extends Migration
             if(!Schema::hasColumn('users','name')){
                 $table->string('name');
             }
-            $table->dropColumn('first_name');
-            $table->dropColumn('last_name');
-            $table->dropColumn('address');
-            $table->dropColumn('city');
-            $table->dropColumn('country');
-            $table->dropColumn('phone');
-            $table->dropColumn('fax');
+            if(Schema::hasColumn('users','first_name')){
+                $table->dropColumn('first_name');
+            }
+            if(Schema::hasColumn('users','last_name')) {
+                $table->dropColumn('last_name');
+            }
+            if(Schema::hasColumn('users','address')) {
+                $table->dropColumn('address');
+            }
+            if(Schema::hasColumn('users','city')) {
+                $table->dropColumn('city');
+            }
+            if(Schema::hasColumn('users','country')) {
+                $table->dropColumn('country');
+            }
+            if(Schema::hasColumn('users','first_name')) {
+                $table->dropColumn('phone');
+            }
+            if(Schema::hasColumn('users','first_name')) {
+                $table->dropColumn('fax');
+            }
         });
     }
 }
